refactor(maestro): split ide/completion/script into cohesive modules
